# lib/api-zod/src/generated/types/batchMapRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MappingConfig } from "./mappingConfig";

export interface BatchMapRequest {
  /**
   * List of entity names to map (max 10,000)
   * @maxItems 10000
   */
  names: string[];
  config?: MappingConfig;
}