# lib/api-zod/src/generated/types/jobResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobResultStatus } from "./jobResultStatus";
import type { MappingResultItem } from "./mappingResultItem";

export interface JobResult {
  job_id: string;
  status: JobResultStatus;
  completed: number;
  total: number;
  error_count: number;
  error_message?: string | null;
  results: MappingResultItem[];
}