# lib/api-zod/src/generated/types/mappingConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MappingConfigAnnotationMode } from "./mappingConfigAnnotationMode";
import type { MappingConfigHints } from "./mappingConfigHints";

export interface MappingConfig {
  annotationMode?: MappingConfigAnnotationMode;
  hints?: MappingConfigHints;
}