# lib/api-zod/src/generated/types/mappingResultItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MappingResultItemConfidenceTier } from "./mappingResultItemConfidenceTier";
import type { MappingResultItemIdentifiers } from "./mappingResultItemIdentifiers";

export interface MappingResultItem {
  name: string;
  resolved: boolean;
  primaryCurie?: string | null;
  confidenceScore?: number | null;
  confidenceTier?: MappingResultItemConfidenceTier;
  needsReview?: boolean;
  identifiers?: MappingResultItemIdentifiers;
  error?: string | null;
  error_type?: string | null;
}